Refactor streamer module, split to separate submodules
