+++ chain/indexer/src/streamer/errors.rs
@@ -1,0 +1,14 @@
+use actix::MailboxError;
+
+/// Error occurs in case of failed data fetch
+#[derive(Debug)]
+pub enum FailedToFetchData {
+    MailboxError(MailboxError),
+    String(String),
+}
+
+impl From<MailboxError> for FailedToFetchData {
+    fn from(actix_error: MailboxError) -> Self {
+        FailedToFetchData::MailboxError(actix_error)
+    }
+}
